20 Mar-For_loop with continue keyword

+++ src/Forloop/Ex_for_loop.java
@@ -2,7 +2,8 @@ package Forloop;
 
 public class Ex_for_loop {
     public static void main(String[] args) {
-        for(int i=0; i<10; i++){
+    //we can use any data type. For float i = 0f
+        for(long i=0l; i<10; i++){
            System.out.println(i);
         }
     }
